Consolidate OI workflow helpers in views
