+++ templates/oi/bits/compare_actions.html
@@ -1,4 +1,3 @@
-{% load editing %}
 <form id="review" action="{% url 'process' id=changeset.id %}" method="POST">
   {% csrf_token %}
   <div>
@@ -7,7 +6,7 @@
   </div>
 {% if user == changeset.indexer %}
   {% if changeset.state == states.OPEN %}
-  <button class="btn-blue-editing inline" type="submit" name="submit">{% workflow_action_label 'submit' %}
+  <button class="btn-blue-editing inline" type="submit" name="submit">{{ workflow_action_labels.submit }}
   </button>
   {% endif %}
   {% if changeset.state == states.REVIEWING or changeset.state == states.OPEN %}
@@ -40,12 +39,12 @@
 
 {% if changeset.state == states.REVIEWING or changeset.state == states.DISCUSSED %}
   {% if user == changeset.approver %}
-  <button class="btn-blue-editing inline" type="submit" name="approve">{% workflow_action_label 'approve' %}</button>
+  <button class="btn-blue-editing inline" type="submit" name="approve">{{ workflow_action_labels.approve }}</button>
     {% if changeset.state != states.DISCUSSED %}
   <button class="btn-blue-editing inline" type="submit" name="discuss">Discuss</button>
     {% endif %}
     {% if changeset.editable %}
-  <button class="btn-blue-editing inline" type="submit" name="disapprove">{% workflow_action_label 'disapprove' %}</button>
+  <button class="btn-blue-editing inline" type="submit" name="disapprove">{{ workflow_action_labels.disapprove }}</button>
     {% endif %}
   <button class="btn-blue-editing inline" type="submit" name="cancel">Reject Permanently</button>
   {% else %}
