+++ templates/oi/edit/changeset.html
@@ -92,7 +92,7 @@
       </td>
     </tr>
   </table>
-  <button class="btn-blue-editing inline" type="submit" name="submit">{% workflow_action_label 'submit' %}</button>
+  <button class="btn-blue-editing inline" type="submit" name="submit">{{ workflow_action_labels.submit }}</button>
       {% if changeset.approver %}
       <button class="btn-blue-editing inline" type="submit" name="discuss">Discuss</button>
       {% endif %}
